requirements.txt
flask
flask-restful
json-schema
flask_restplus
cachetools
sqlalchemy
flask-sqlalchemy
flask_marshmallow
marshmallow-sqlalchemy
Faker
python-dotenv

#
pytest

